refactor: move metadata defaults into RFC module
